Rewrites server routes to send queries to either database
    - Updates README to CRUD API version
