+++ README.md
@@ -39,3 +39,108 @@ npm install -g webpack
 npm install
 ```
 
+Cassandra CRUD API
+------
+
+Create/POST
+------
+'/restaurants/:id/order'
+```
+INSERT INTO orders (name, item_id, restaurant_id, choices, quantity, special, total)
+    VALUES (?, ?, ?, ?, ?, ?, ?)
+```
+
+'/restaurants/:id/menu-items'
+```
+INSERT INTO menu3 (
+    restaurant_id, item_id, name, 
+    price, picture_url, category, 
+    description, required, optional
+) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?);
+```
+
+Read/GET
+------
+'/restaurants/:id/menu-items'
+```
+SELECT * FROM menu WHERE restaurant_id = ?
+```
+'/restaurants/:id/menu-items/:itemId'
+```
+SELECT * FROM menu WHERE restaurant_id = ? AND item_id = ?
+```
+
+Update/PUT
+------
+'/restaurants/:id/menu-items/:itemId'
+```
+UPDATE menu SET (
+    name, price, picture_url, category, 
+    description, required, optional
+) WHERE restaurant_id = ? AND item_id = ?
+```
+
+Delete/DELETE
+------
+'/restaurants/:id/menu-items/:itemId'
+```
+DELETE FROM menu WHERE restaurant_id = ? AND item_id = ?
+```
+
+PostgreSQL CRUD API
+=
+___
+
+Create/POST
+-
+'/restaurants/:id/order'
+  ```
+  INSERT INTO orders (
+    name, item_id, restaurant_id, 
+    choices, quantity, special, total
+  ) VALUES ($, $, $, $, $, $, $);
+  ```
+  
+'/restaurants/:id/menu-items'
+  ```
+  INSERT INTO menu (
+    item_id, restaurant_id, name, 
+    price, picture_url, category, 
+    description, required, optional
+  ) VALUES ($, $, $, $, $, $, $, $, $);
+  ``` 
+
+Read/GET
+-
+'/restaurants/:id/menu-items'
+  ```
+  SELECT * FROM menu WHERE restaurant_id = $;
+  ```
+
+'/restaurants/:id/menu-items/:itemId'
+  ```
+  SELECT * FROM menu 
+    WHERE restaurant_id = $ AND item_id = $;
+  ```
+  
+Update/PUT
+-
+'/restaurants/:id/menu-items/:itemId'
+  ```
+    UPDATE menu SET 
+      name = $1,
+      price = $2,
+      picture_url = $3,
+      category = $4,
+      description = $5,
+      required = $6,
+      optional $7
+    WHERE restaurant_id = $ AND item_id = $;
+  ```
+  
+Delete/DELETE
+-
+'/restaurants/:id/menu-items/:itemId'
+  ```
+  DELETE FROM menu WHERE restaurant_id = $ AND item_id = $;
+  ```
