+++ database/postgres.js
@@ -1,21 +1,35 @@
 const { Client } = require('pg');
+
 const client = new Client({
-    user: 'sdc',
-    host: 'localhost',
-    database: 'sdc'
+  user: 'sdc',
+  host: 'localhost',
+  database: 'sdc',
 });
 client.connect();
 
-async function res() {
-    const r = await client.query('SELECT NOW()');
-    console.log(r.rows);
-    return r;
-}
+const getMenu = async (params) => {
+  const query = 'SELECT * FROM menu WHERE restaurant_id=$1';
+  try {
+    const menu = await client.query(query, params);
+    return menu.rows;
+  } catch (err) {
+    console.log('ERROR GETTING MENU', err);
+    return err;
+  }
+};
 
-async function sample() {
-    const r = await client.query('SELECT name FROM menu WHERE restaurant_id = 3 AND item_id = 2');
-    console.log(r.rows);
-}
+const getItem = async (params) => {
+  const query = 'SELECT * FROM menu WHERE restaurant_id=$1 AND item_id=$2';
+  try {
+    const item = await client.query(query, params);
+    return item.rows;
+  } catch (err) {
+    console.log('ERROR GETTING ITEM', err);
+    return err;
+  }
+};
 
-module.exports.res = res;
-module.exports.sample = sample;
+module.exports = {
+  getMenu,
+  getItem,
+};
